PERA-1812 :: move settings ui files to ui/settings folder
